microservices/lambda-function/src/requirements.txt
boto3
gitpython
PyGithub